+++ frontend/src/ancestors/DeleteAncestor/AncestorDeletion.css
