+++ frontend/src/ancestors/DeleteAncestor/AncestorDeletion.tsx
