--- frontend/src/components/SingleFileUploader.css
@@ -1,32 +1,0 @@
-.conteneur-détails {
-    display: flex;
-    padding: 10px 30px 20px 70px;
-    background-color: rgba(255, 0, 0, 0.349);
-}
-
-ul {
-    display: flex;
-    flex-direction: column;
-    list-style-type: none;
-    margin: 5px;
-    padding: 8px;
-    overflow: hidden;
-    background-color: var(--primary-dark-green);
-    border:hsl(from color h s l);
-    border-radius: 4px;
-    color: white;
-    font-size: 0.8rem;
-}
-
-li {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin: 2px;
-    padding: 2px;
-}
-
-.bouton-parcourir {
-    padding: 10px;
-    margin: 10px;
-}
