+++ frontend/src/components/SingleFileUploader/SingleFileUploader.css
@@ -1,0 +1,150 @@
+.conteneur-détails {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 5px 30px 10px;
+    background-color: rgba(200, 255, 0, 0.349);
+    position: relative;
+}
+
+.truc {
+    display: flex;
+    flex-direction: row;
+    background-color: white;
+    position: relative;
+}
+
+.closing-X {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    align-self: end;
+    margin: 5px;
+    height: 20px;
+    width: 20px;
+    background-color: transparent blue;
+    opacity: 50%;
+    border: 4px solid grey;
+    border-radius: 15px;
+    color: black;
+    font-size: 18px;
+    font-weight: bolder;
+    cursor: pointer;
+    /* position: absolute; */
+    /* top: 0;
+    right: 8px; */
+
+    /*Prévoir style quand le bouton est pressé/non pressé*/
+}
+
+h4 {
+    text-align: center;
+    width: 100%;
+    margin: 5px;
+}
+
+ul {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    list-style-type: none;
+    margin: 2px;
+    padding: 2px;
+    flex-grow: 1;
+    overflow: hidden;
+    background-color: var(--primary-dark-green);
+    border: hsl(from color h s l);
+    border-radius: 4px;
+    color: white;
+    font-size: 0.8rem;
+}
+
+li {
+    background-color: lightseagreen;
+    display: flex;
+    justify-content: start;
+    align-items: center;
+    margin: 0.5px;
+    padding: 0.5px;
+    border-radius: 4px;
+    font-size: 11px;
+}
+
+.status-messages>span {
+    margin: 2px;
+    padding: 2px;
+    border: 2px solid var(--primary-dark-blue);
+    border-radius: 5px;
+}
+
+.status-messages span:nth-child(1) {
+    background-color: lightgreen;
+}
+
+.status-messages span:nth-child(2) {
+    background-color: lightcoral;
+}
+
+.status-messages span:nth-child(3) {
+    background-color: lightyellow;
+}
+
+
+.miniature {
+    display: flex;
+    margin: 5px;
+    /*Déplacer style miniature ici - Pour l'instant présent inline pour débuggage.*/
+}
+
+.miniature:hover {
+    border: 6px solid var(--primary-dark-blue);
+    border-radius: 35px;
+    width: 200px;
+    height: 200px;
+    cursor: pointer;
+}
+
+/* .bouton-parcourir {
+    padding: 5px;
+    margin: 10px;
+    text-align: center;
+    color: red;
+} */
+
+
+/*--------------------*/
+
+
+.file-input {
+    background-color: yellow;
+    /* display: none; */
+    width: auto;
+    margin-left: 10px;
+}
+
+.file-label {
+    margin-right: 50px;
+    color: grey;
+    font-size: 14px;
+    /* display: none; */
+}
+
+.file-label:hover {
+    color: black;
+    cursor: pointer;
+}
+
+.custom-button {
+    padding: 8px 16px;
+    background-color: #4CAF50;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+}
+
+.custom-button:hover {
+    background-color: #45a049;
+}
